Move the time.Sleep() to after DB instantiation
Based on my investigation, it seems PouchDB does an unexplained setTimeout
after instantiating a DB using indexedDB, so I'm thinking this bug may
happen for any operation against indexeddb, not just purge.  So best to do
our noop sleep then.

Although it would be great to truly understand this, and remove it if/when
possible.